+++ CHANGELOG.md
@@ -12,6 +12,7 @@ MAGE adheres to [Semantic Versioning](http://semver.org/).
 * Fix kml to geojson icon style.
 * Export observations based on timestamp, not last updated.
 * CSV export properly handles commas in values.
+* Fix permission error when event user posts recent event.
 
 ## [5.2.6](https://github.com/ngageoint/mage-server/releases/tag/5.2.6) (TBD)